Implement conditional autoprompt importing.
