+++ tests/unit/autoprompt/test_core.py
@@ -28,70 +28,6 @@ class TestCLIAutoPrompt(unittest.TestCase):
             self.driver, prompter=self.prompter
         )
 
-    def test_throw_error_if_both_args_specified(self):
-        args = ['--cli-auto-prompt', '--no-cli-auto-prompt']
-        self.assertRaises(
-            ParamValidationError,
-            self.prompt_driver.validate_auto_prompt_args_are_mutually_exclusive,
-            args,
-        )
-
-
-def _generate_auto_prompt_resolve_cases():
-    # Each case is a 5-namedtuple with the following meaning:
-    # "args" is a list of arguments that command got as input from
-    #   command line
-    # "config_variable" is the result from get_config_variable
-    #   This takes a value of either 'on' , 'off' or 'on-partial'
-    # "expected_result" is a boolean indicating whether auto-prompt
-    #   should be used or not.
-    #
-    # Note: This set of tests assumes that only one of --no-cli-auto-prompt
-    # or --cli-auto-prompt overrides can be specified.
-    # TestCLIAutoPrompt.test_throw_error_if_both_args_specified tests
-    # that these command line overrides are mutually exclusive.
-    Case = namedtuple(
-        'Case',
-        [
-            'args',
-            'config_variable',
-            'expected_result',
-        ],
-    )
-    return [
-        Case([], 'off', 'off'),
-        Case([], 'on', 'on'),
-        Case(['--cli-auto-prompt'], 'off', 'on'),
-        Case(['--cli-auto-prompt'], 'on', 'on'),
-        Case(['--no-cli-auto-prompt'], 'off', 'off'),
-        Case(['--no-cli-auto-prompt'], 'on', 'off'),
-        Case([], 'on', 'on'),
-        Case([], 'on-partial', 'on-partial'),
-        Case(['--cli-auto-prompt'], 'on-partial', 'on'),
-        Case(['--no-cli-auto-prompt'], 'on-partial', 'off'),
-        Case(['--version'], 'on', 'off'),
-        Case(['help'], 'on', 'off'),
-    ]
-
-
-@pytest.mark.parametrize('case', _generate_auto_prompt_resolve_cases())
-def test_auto_prompt_resolve_mode(case):
-    driver = create_clidriver()
-    driver.session.set_config_variable('cli_auto_prompt', case.config_variable)
-    prompter = mock.Mock(spec=core.AutoPrompter)
-    prompt_driver = core.AutoPromptDriver(driver, prompter=prompter)
-    result = prompt_driver.resolve_mode(args=case.args)
-    assert result == case.expected_result
-
-
-def test_auto_prompt_resolve_mode_on_non_existing_profile():
-    driver = create_clidriver()
-    driver.session.set_config_variable('profile', 'not_exist')
-    prompter = mock.Mock(spec=core.AutoPrompter)
-    prompt_driver = core.AutoPromptDriver(driver, prompter=prompter)
-    result = prompt_driver.resolve_mode(args=[])
-    assert result == 'off'
-
 
 class TestAutoPrompter(unittest.TestCase):
     def setUp(self):
